requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
pdfplumber>=0.11.4
spacy>=3.8.0
reportlab>=4.2.2
PyYAML>=6.0.2
